requirements.txt
Django==5.2.5
djangorestframework==3.15.0
dj-config-url==0.1.1
sqlparse==0.5.3
asgiref==3.9.1
tzdata==2025.2

# DB driver (
psycopg2-binary>=2.9
# Testing
pytest==8.4.1
pytest-django==4.11.1

# Dev tools
black==25.1.0
ruff==0.12.9

dj-database-url==2.3.0
psycopg2-binary==2.9.9
